Create doc version system
each document type has a package-dir, like libros, and each class refers to a version, like v9.py. So, we can use doctype.libros.v9.LibrosDoc. In the doctype __init__ we adapt the enum to use both DOCTYPE and VERSION fields in order to choose a class to parse the rest of the doc