Refactor how sinli codes are defined and used
Create a class for each sinli code instead of a dictionary.
This way we can go back and forth between encoded sinli value and
coder-usable identifier. We can translate to a human-readable description too.

Now, line fields of some sinli code type, are objects subclassed of
FieldCodes, and are therefore comparable with readable names, that is
"NORMAL" instead of the ambiguous "N"